requirements.txt
bidict==0.23.1
certifi==2025.10.5
charset-normalizer==3.4.3
colorama==0.4.6
dnspython==2.7.0
eventlet==0.40.3
greenlet==3.2.4
h11==0.16.0
idna==3.10
loguru==0.7.3
python-engineio==4.12.3
python-socketio==5.14.1
requests==2.32.5
rti.connext==7.5.0.1
rticonnector @ file:///C:/Projects/rticonnector-1.0.10-py3-none-any.whl
shortio==0.1.0
simple-websocket==1.1.0
sio==0.1.0
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
win32_setctime==1.2.0
wsproto==1.2.0
